tasks/requirements.txt
cookiecutter==1.6.0
future==0.16.0

invoke==1.1.1
# 3.11 was incompatible with kaptan for git-autoshare
PyYAML==6.0
requests==2.20.0
ruamel.yaml>=0.15.66
gnupg==2.3.1
psycopg2-binary>=2.7.6
python-dateutil==2.7.3
git-aggregator >= 1.7.2

git-autoshare >= 1.0.0b2
passlib

clipboard==0.0.4

marabunta >= 0.9.0

gitpython >= 3.1.11
pre-commit
